requirements.txt
streamlit
langchain
langchain-community
langchain-openai
chromadb
pypdf
python-dotenv
langchain-text-splitters
